Refactored SpriteSheet into seperate package, made SpriteFrame external class
Created cs410.lanbros.animation package, moved SpriteSheet and SpriteFrame to package
Added restartAnimation, offsetAnimation, getCurrentFrame, getCurrentAnimation to SpriteSheet
When unpausing a manually frozen SpriteSheet, will resume animating
Moved SpriteFrame outside of SpriteSheet
